Add nested output mode
We try to translate nested selectors into the appropriate set of nested
lists of styles with 'Global' helpers where needed.

There falls back to use UnknownNode and UnknownEntry to handle some
awkward points but otherwise produces syntactically correct Elm code
with some promise.

Not saying that the code is the cleanest though.