Use descendants to fix compile errors
Previously we weren't using anything like descendants which meant that
we had 'Snippets' in our 'Style' lists which the compiler didn't like.

This is all still a bit hacky but we say that if the first letter of a
nested selector isn't an '&' then we must want a 'descendants'
relationship.

This finally produces compiling code for a small production css file.